Moving all vector store interfaces to data folder and namespace.
