+++ dotnet/src/Connectors/Connectors.Memory.Redis/RedisVectorStoreRecordMapper.cs
@@ -2,7 +2,7 @@
 
 using System.Text.Json;
 using System.Text.Json.Nodes;
-using Microsoft.SemanticKernel.Memory;
+using Microsoft.SemanticKernel.Data;
 
 namespace Microsoft.SemanticKernel.Connectors.Redis;
