+++ dotnet/src/InternalUtilities/src/Data/VectorStoreErrorHandler.cs
@@ -3,9 +3,8 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
-using Microsoft.SemanticKernel.Memory;
 
-namespace Microsoft.SemanticKernel;
+namespace Microsoft.SemanticKernel.Data;
 
 /// <summary>
 /// Contains helpers for reading vector store model properties and their attributes.
