+++ dotnet/src/SemanticKernel.Abstractions/Data/VectorStoreRecordMappingException.cs
@@ -3,7 +3,7 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 
-namespace Microsoft.SemanticKernel.Memory;
+namespace Microsoft.SemanticKernel.Data;
 
 /// <summary>
 /// Exception thrown when a failure occurs while trying to convert models for storage or retrieval.
